requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-chroma
chromadb
openai
pdfplumber
pdfminer.six
python-dotenv
tqdm
protobuf==3.20.3
pysqlite3-binary
retrying